+++ lib/sequence.dart
@@ -6,7 +6,6 @@ part 'src/sequence/delimited.dart';
 part 'src/sequence/pair.dart';
 part 'src/sequence/preceded.dart';
 part 'src/sequence/separated_pair.dart';
-part 'src/sequence/skip.dart';
-part 'src/sequence/skip_many0.dart';
+part 'src/sequence/sequence.dart';
 part 'src/sequence/terminated.dart';
 part 'src/sequence/tuple.dart';